Fixed linux build
--HG--
branch : sourcemod-1.0.x
extra : convert_revision : svn%3A39bc706e-5318-0410-9160-8a85361fbb7c/branches/sourcemod-1.0.x%402084

+++ extensions/tf2/criticals.cpp
@@ -203,7 +203,7 @@ bool CriticalHitManager::CreateCriticalMeleeDetour()
 	/* Patch old bytes in */
 	for (size_t i=0; i<melee_restore.bytes; i++)
 	{
-		if (i != callbyte)
+		if ((int)i != callbyte)
 		{
 			jit->write_ubyte(melee_restore.patch[i]);
 			continue;
